feat: updated freezed models
- separated abilities,moves, and stats detail classes
