+++ pokedex_async_redux/lib/api/model/models.dart
@@ -1,5 +1,5 @@
 export 'pokemon.dart';
-export 'pokemon_moves.dart';
-export 'pokemon_abilities.dart';
-export 'pokemon_details.dart';
-export 'pokemon_stats.dart';
+export 'moves.dart';
+export 'ability.dart';
+export 'details.dart';
+export 'stats.dart';
