+++ pokedex_async_redux/lib/api/model/moves_details.dart
@@ -1,16 +1,7 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-part 'pokemon_moves.freezed.dart';
-part 'pokemon_moves.g.dart';
-
-@freezed
-class PokemonMoves with _$PokemonMoves {
-  const factory PokemonMoves({
-    @JsonKey(name: 'move') required MovesDetails move,
-  }) = _PokemonMove;
-
-  factory PokemonMoves.fromJson(Map<String, dynamic> json) => _$PokemonMovesFromJson(json);
-}
+part 'moves_details.freezed.dart';
+part 'moves_details.g.dart';
 
 @freezed
 class MovesDetails with _$MovesDetails {
