+++ pokedex_async_redux/lib/api/model/pokemon.dart
@@ -7,8 +7,8 @@ part 'pokemon.g.dart';
 @freezed
 class Pokemon with _$Pokemon {
   factory Pokemon({
-    @JsonKey(name: 'name') String name,
-    @JsonKey(name: 'url') String url,
+    @Default("") @JsonKey(name: 'name') String name,
+    @Default("") @JsonKey(name: 'url') String url,
   }) = _Pokemon;
 
   factory Pokemon.fromJson(Map<String, dynamic> json) => _$PokemonFromJson(json);
