+++ pokedex_async_redux/lib/main.dart
@@ -26,5 +26,3 @@ Future<void> main() async {
     ),
   );
 }
-
-//test
